Update celltypes.py
Came across issues with the regex, that caused mapping gene names to fail.

Also, the issue with ValueError: DataFrame constructor not properly called occurred due to the inner function, _guess_cell_type, returning a series rather than a list of dicts. Converting that to a list solved the problem for me.